Move `forward_dynamics` and `nonlinear_state_space` functions `systems/utils` and remove `substitute_params_into_all_symbolic_expressions` and `substitute_params_into_single_symbolic_expression`
